+++ triplea/cli/config.py
@@ -3,5 +3,6 @@ from triplea.cli.main import cli
 
 @cli.command('config',help = 'Configuration additional setting. ')
 def configuration():
+    raise NotImplementedError
     pass